requirements.txt
Pillow
streamlit
fastapi[all]
google-genai
python-dotenv
python-multipart